refactor: rename phase traces components directory
