+++ src/component/2d/1d-tracer/phase-correction-traces/SpectraPhaseTraces.tsx
